+++ CHANGELOG.mdx
@@ -88,6 +88,7 @@ Prefix the change with one of these keywords:
 - `Figure`: replaced primary div wrapper with `Figure` component
 - `Login`: change top margin to fit logo inside container
 - `Modal`: removed section wrapper, updated close button styles
+- `Table`: remove duplicate `cu-table` class
 
 ### Fixed
